Create character_to_binar_hexadecimal.c
Write a C program to create an encrypted code of a given string. The encrypted code is either in binary number or in hexadecimal number. Rules to create the encrypted code:

    You have to jumble alphabets of the given string in such a way that alphabets in the odd positions come together first and alphabets in even positions come together after that. For example, if the string is APPLE, the jumbled string according to the rule given above is APEPL.
    You have to find the binary/hexadecimal code of the jumbled string and print it.

Use switch-case for implementing the menu driven program. The programmer is supposed to use the indicated functions for the choices, case 1: string_to_binary(); case 2: string_to_hex().

So in the above mentioned example, corresponding to APEPL, case 1 will give you binary number 0100000101010000010001010101000001001100; and case 2 will give you hexadecimal number 415045504c.